test: mock useSoftwareList on test
