+++ __tests__/test-utils.ts
@@ -5,6 +5,7 @@ import flushPromises from 'flush-promises'
 import type { Plugin } from 'vue'
 import { defineComponent, h, Suspense } from 'vue'
 
+import type { SoftwareListData } from '~~/composables/useSoftwareList'
 import type { SoftwareParsedContent } from '~~/src/content'
 
 export const plugins: (Plugin | [Plugin, ...any[]])[] = [[Oruga, bulmaConfig]]
@@ -47,3 +48,20 @@ export const mockSoftware: SoftwareParsedContent = {
   body: '',
   _id: 'content:1st-jp:index.md'
 }
+
+export const mockSoftwareList: SoftwareListData[] = [
+  {
+    slug: '1st-jp',
+    title: 'Dance Dance Revolution',
+    platform: 'Play Station',
+    region: 'JP',
+    launched: '1999-04-10'
+  },
+  {
+    slug: '2nd',
+    title: 'Dance Dance Revolution 2nd ReMIX',
+    platform: 'Play Station',
+    region: 'JP',
+    launched: '1999-08-26'
+  }
+]
